refactor: extract studio clock specific logic
