+++ apps/client/src/features/viewers/studio/__tests__/studioClock.utils.test.ts
@@ -1,4 +1,4 @@
-import { formatEventList, trimRundown } from '../eventsManager';
+import { formatEventList, trimRundown } from '../studioClock.utils';
 
 describe('test trimEventlist function', () => {
   const limit = 8;
